refactor: 토스트 타입 정의 파일명 변경
