+++ packages/design-system/src/components/toast/utils/eventManager.ts
@@ -1,4 +1,4 @@
-import { EventManager } from '../types/type';
+import { EventManager } from '../types';
 
 export const eventManager: EventManager = {
   list: new Map(),
